+++ visualization/app/codeCharta/codeCharta.model.ts
@@ -95,6 +95,7 @@ export interface FileMeta {
 	apiVersion: string
 	projectName: string
 	exportedFileSize: number
+	repoCreationDate?: string
 }
 
 export interface Settings {