+++ visualization/app/codeCharta/util/dataMocks.ts
@@ -719,7 +719,7 @@ export const FILE_META: FileMeta = {
 	fileChecksum: "md5-fileA",
 	projectName: "Sample Project",
 	apiVersion: packageJson.codecharta.apiVersion,
-	exportedFileSize: 300000
+	exportedFileSize: 300_000
 }
 
 export const TEST_FILE_DATA: CCFile = {
@@ -1323,7 +1323,7 @@ export const TEST_DELTA_MAP_A: CCFile = {
 		fileChecksum: "md5-delta-fileA",
 		projectName: "Sample Project",
 		apiVersion: packageJson.codecharta.apiVersion,
-		exportedFileSize: 300000
+		exportedFileSize: 300_000
 	},
 	map: {
 		name: "root",
@@ -1381,7 +1381,7 @@ export const TEST_DELTA_MAP_B: CCFile = {
 		fileChecksum: "md5-delta-fileB",
 		projectName: "Sample Project",
 		apiVersion: packageJson.codecharta.apiVersion,
-		exportedFileSize: 300000
+		exportedFileSize: 300_000
 	},
 	map: {
 		name: "root",
@@ -1455,7 +1455,7 @@ export const TEST_DELTA_MAP_C: CCFile = {
 		fileChecksum: "md5-delta-fileB",
 		projectName: "Sample Project",
 		apiVersion: packageJson.codecharta.apiVersion,
-		exportedFileSize: 300000
+		exportedFileSize: 300_000
 	},
 	map: {
 		name: "root",
@@ -1529,7 +1529,7 @@ export const TEST_DELTA_MAP_D: CCFile = {
 		fileChecksum: "md5-delta-fileB",
 		projectName: "Sample Project",
 		apiVersion: packageJson.codecharta.apiVersion,
-		exportedFileSize: 300000
+		exportedFileSize: 300_000
 	},
 	map: {
 		name: "root",