+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 ### Fixed 🐞
 
 -   Empty temporary label during hovering ([#2328](https://github.com/maibornwolff/codecharta/issues/2328))
+-   Show the screenshot hotkey in the screenshot title ([#2323](https://github.com/maibornwolff/codecharta/issues/2323))
 
 ## [1.77.0] - 2021-07-30